requirements.txt
Flask==2.3.3
Flask-Cors==3.1.3
SQLAlchemy==2.0.21
psycopg2-binary==2.9.6
requests==2.31.0
python-dotenv==1.0.1
google-generativeai==0.6.0
